Lazily execute formatting of custom block
This allow user to provide custom format without having to handle
themselves the order of execution of the dependencies.

For instance, one could now write:

  {
    // In slack message

    git {
      format {
        if (project.version == "42") {
          section {
            // ...
          }
        }
      }
    }
  }

  // After
  project.version = "42"

Before, the format was eagerly executed, and code inside the slack
message directives were lazily executed. This would result, at best, to
incomprehension from user, and, at worst, to weird bug when same
property is used in and out the directives.